Issue 207: Fix silently ignoring documents under load
This fixes an assumption in the indexer where it was assumed that if a
pending document had an ID, all the subsequent items in the chain had
IDs as well. This assumption will fail if between the merging of entries
and the assignment of document IDs, more requests are added to the
chain; resulting in entries which are unmerged but have IDs. In such a
case, during the next iteration of Indexer_Process(), the function would
mistakenly assume that all subsequent IDs in the chain had been assigned.